+++ changelog.md
@@ -1,4 +1,8 @@
 # Changelog - Amazon Price Tracker
+## November 28, 2019
+* Cleaned up the code & split classes into separate files
+* Added feature to remove items from tracking list
+* Items are now automatically removed once they drop below their target price and alerts are sent out
 ## November 26, 2019
 * Now creating changelogs!
 * Fixed an issue with commas in item names